# swagger.yml
swagger: '2.0'
info:
  description: Conduit API
  version: 1.0.0
  title: Conduit API
  contact:
    name: RealWorld
    url: 'https://realworld.io'
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
basePath: /api
host: localhost:8080
schemes:
  # - https
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Token:
    description: |
      For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.

      A JWT token is generated by the API by either registering via /users or logging in via /users/login.

      The following format must be in the 'Authorization' header :

          Token: xxxxxx.yyyyyyy.zzzzzz

    type: apiKey
    name: Authorization
    in: header
paths:
...
  /articles:
    get:
      summary: Get recent articles globally
      description: Get most recent articles globally. Use query parameters to filter results. Auth is optional
      tags:
        - Articles
      operationId: GetArticles
      parameters:
        - name: tag
          in: query
          description: Filter by tag
          required: false
          type: string
        - name: author
          in: query
          description: Filter by author (username)
          required: false
          type: string
        - name: favorited
          in: query
          description: Filter by favorites of a user (username)
          required: false
          type: string
        - name: limit
          in: query
          description: Limit number of articles returned (default is 20)
          required: false
          default: 20
          type: integer
        - name: offset
          in: query
          description: Offset/skip number of articles (default is 0)
          required: false
          default: 0
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MultipleArticlesResponse'
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
          schema:
            $ref: '#/definitions/GenericErrorModel'
    post:
      summary: Create an article
      description: Create an article. Auth is required
      tags:
        - Articles
      security:
        - Token: []
      operationId: CreateArticle
      parameters:
        - name: article
          in: body
          required: true
          description: Article to create
          schema:
            $ref: '#/definitions/NewArticleRequest'
      responses:
        '201':
          description: OK
          schema:
            $ref: '#/definitions/SingleArticleResponse'
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
          schema:
            $ref: '#/definitions/GenericErrorModel'
...
definitions:
...
  Profile:
    type: object
    properties:
      username:
        type: string
      bio:
        type: string
      image:
        type: string
      following:
        type: boolean
    required:
      - username
      - bio
      - image
      - following
  Article:
    type: object
    properties:
      slug:
        type: string
      title:
        type: string
      description:
        type: string
      body:
        type: string
      tagList:
        type: array
        items:
          type: string
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
      favorited:
        type: boolean
      favoritesCount:
        type: integer
      author:
        $ref: '#/definitions/Profile'
    required:
      - slug
      - title
      - description
      - body
      - tagList
      - createdAt
      - updatedAt
      - favorited
      - favoritesCount
      - author
  SingleArticleResponse:
    type: object
    properties:
      article:
        $ref: '#/definitions/Article'
    required:
      - article
  MultipleArticlesResponse:
    type: object
    properties:
      articles:
        type: array
        items:
          $ref: '#/definitions/Article'
      articlesCount:
        type: integer
    required:
      - articles
      - articlesCount
  NewArticle:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
      body:
        type: string
      tagList:
        type: array
        items:
          type: string
    required:
      - title
      - description
      - body
  NewArticleRequest:
    type: object
    properties:
      article:
        $ref: '#/definitions/NewArticle'
    required:
      - article
...
  GenericErrorModel:
    type: object
    properties:
      errors:
        type: object
        properties:
          body:
            type: array
            items:
              type: string
        required:
          - body
    required:
      - errors